requirements.txt
Django==3.2.18
psycopg2==2.9.5
djangorestframework==3.14.0
drf-yasg==1.21.5